HERO

+++ desktop/src/com/ever/ending/desktop/DesktopLauncher.java
@@ -8,8 +8,8 @@ import com.ever.ending.management.GameConstants;
 public class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		config.width = GameConstants.SCREEN_WIDTH;
-		config.height = GameConstants.SCREEN_HEIGHT;
+		config.width = (int)(GameConstants.SCREEN_WIDTH);
+		config.height = (int)(GameConstants.SCREEN_HEIGHT);
 		config.fullscreen = false;
 		new LwjglApplication(new EverEnding(), config);
 	}
